MediaWiki: Don't eagerly initialize action name
main() calls getAction() eagerly as it may be used by the file cache
if it is enabled. However, getAction() is expensive as it will trigger
memcached I/O and DB queries, and the file cache is disabled by default
(and by large installations such as Wikimedia). This can result in
unnecessary I/O on requests that abort early due to
tryNormaliseRedirect() or a MediaWikiPerformAction / BeforeInitialize
hook handler.

So, call getAction() early only if the file cache is actually in use.
This avoids the upfront cost of initializing a full WikiPage object
for requests that will not use it.

Change-Id: I72ffc9f36613bf9de0b3d95276435d08a8a65df4