Fire a change event on the checkboxes when selecting All/None/Invert
Fire a change event through JQuery when the user selects one of
All / Invert / None when using the list toggle provided in
ListToggle.php so that any JS that listens for a change event on
the checkboxes (that would be for the user selecting or de-
selecting checkboxes) gets a change event.

This is needed because CheckUser's caMultiLock.js script listens
for change events on the checkboxes in the 'Get edits' results
to update the URL that links to the MultiLock special page.

Bug: T311924
Change-Id: I76af8f2bbd6707dacef9872bb5a737823fc3d0af