Language::getHumanTimestamp: Don't show time when timestamp in a different year
We don't show the time for timestamps that are more than 5 days old
elsewhere in this function, so we shouldn't start showing them again
for timestamps in another year.

Keep the full timestamp format for timestamps in the future, with the
TODO to improve it.

Improve comments throughout the function to be more accurate.

Bug: T133468
Change-Id: I68c6105cc83ca12e0a8f41e8b611bf4b8f35cd87