user: Move logic for unregistered HTTPS pref to User::requiresHTTPS()
Move the User::isRegistered call into the function to let all callers
benefit from that check. This is more than an error check, it also is
business logic representing the effective default for unregistered users
which is non-obvious and benefits from being written down as part of
a central contract.

Update tests as the user must exists on the database to have
preferences.

Change-Id: Ibacdd642d03031d9659a821045788046f0187518