RELEASE-NOTES: dynamic defaults applied before extension registration.
Dynamic default values are now applied before extension registration callbacks are run.
This way, extensions have a complete view of config variables, with all defaults applied.
For example, when the default value of X used to be static but becomes dynamic,
and an extension reads the value of X in the registration callback, it will now continue
to function as expected. In some cases however, this may cause an undesired change
in behavior: if the dynamic default of setting X depends on the value of setting Y,
and an extension changes Y, the changed value of Y will no longer affect the value of X.

Bug: T305093
Change-Id: I69d4167a6302fc3b2d71c8d27b84264d9021fb09