Don't throw away user's text when rate limit is exceeded
When a page save fails due to a rate limit being exceeded, show an error
message and the edit form with the submitted text, instead of throwing
away the submitted text.

Most error display in EditPage is done in showHeader(), hence the lack
of examples to follow here in handleStatus(). PermissionsError has a
bunch of useful logic so is not easily replaced, and permissions errors
may also be shown before save. ReadOnlyError handling is duplicated in
edit() and showHeader() so the exception here is usually not thrown.

Change-Id: I7cb268fb6e20888cddef73d0ee9b0bf060ccafda