objectcache: Accelerate async memcached writes with "noreply"
Instead of non-blocking I/O, use the "noreply" option instead.

Per benchmark results and code investigation at T310662,
switching off Memcached::OPT_BUFFER_WRITES or Memcached::OPT_NO_BLOCK
causes libmemcached to drop any open connections, incurring a latency
penalty when such methods are called in a loop.

So, use the "noreply" option instead. Without OPT_NO_BLOCK, this is
becomes a single sendto() syscall, and it doesn't block until the TCP
window fills up.

Theoretically OPT_NO_BLOCK could be faster if a setMulti() fills up the
TCP window for many servers, since in blocking mode the client might
not be able to keep all the TCP buffers full. But it would have to be
used with a dedicated connection, so it's an optimisation only suited
for a special case.

Change-Id: I11e3b9394821fd6dcc5c1adb42bb546659ed5af8