migrateLinksTable.php: BETWEEN is inclusive
See, for example, MariaDB documentation:

https://mariadb.com/kb/en/between-and/

In part because of this, the code was including the first two page IDs
of an outer batch in the previous outer batch as well. Fortunately, this
was the case as opposed to failing to update those rows. In any case,
this commit corrects the calculation of $highPageId so the batching
works as expected.

Change-Id: Icb5eb5797d7a6cf992895224e091aad1da499724