migrateLinksTable.php: Remove LIMIT from UPDATE query
On a live site configured for WRITE_NEW, no new templatelinks rows in
which tl_target_id IS NULL should be added. Assuming this is the case,
the conditions on PK fields (tl_from, tl_namespace, tl_title) should
suffice to limit the number of affected rows, since the range of page
IDs is limited and only one distinct target ID is filled at a time.

This avoids a needless "Updated 0 rows" query for every link target
within a batch.

Change-Id: I8af72ca8fe4c68da1603670d065a2b51ab96b79d