EditPage: Remove inconsistent "Preview of subject:" label when adding new sections
Despite the label stating "Preview of subject:", the thing shown
afterwards was actually the preview of the entire edit summary.
This has been the case since 2008 (b49d9f5ae3).

Previewing just the subject (section title) would be useless, since
it's included in the main preview.

Also fix live preview to preview the whole summary, not just the subject.

Change-Id: I9c3acb2e722d70ffe57bdbaa7cc175ef33e91928