Introduce authorizeAction() into Authority.
In order to perform all rate limit checks through Authority, we need
methods that authorize an action in the context of the current state
of the system. We already have userCan, authorizeRead and
authorizeWrite that do this, but they requrie a specific page for
context. We need to be able to do the same for actions that do not
apply to a specific page (e.g. resetting a user's password).

This patch fills the gap by introducing authorizeAction() and
isDefinitelyAllowed() for actions that do not apply to a specific
page.

This change makes use of the new methods in a couple of places to
illustrate the idea:
- ChangeTags uses authorizeAction instead of User::pingLimiter

As an aside, this moves newFatalPermissionDeniedStatus from the
User class into PermissionManager, to avoid the use of global state
and thus allow UserAuthorityTest to function without having to
initialize MediaWiki.

Bug: T310476
Change-Id: I0c2ce8bdf079c6892c6ffd6ddf0baf882b2adcb7