Ensure custom locales for Moment.js overrides, don't change 'en'
Define a locale if the current language doesn't have one, so that
we can apply overrides without affecting the defaults for English.

We need this because code in various places expects that
`moment( … ).locale( 'en' ).format( … )` will return results
with non-localized digits.
https://codesearch.wmcloud.org/search/?q=%5C.locale%5C(%20%27en%27%20%5C)

This can be tested most easily by checking that it's possible to pick
a date in the date dropdowns on Special:Contributions?uselang=ckb.

Bug: T313188
Change-Id: Ic82064fe29febb950ebcedad9a6081ad651081bc