Move knowledge about the attribs array out of TransformContext
This renames TransformContext to HTMLTransformInput. It is becoming a
wrapper around the input HTML, with a bunch of optional context data
attached.

This introduces a factory method for HTMLTransformInput, so we can
extract knowledge about the structure of the $attribs array from
HTMLTransformInput.

This also allows us to inject Document objects and perhaps PageBundle
objects, instead of just arrays.

Change-Id: I66f9c5dbb50c6bf1f582adad7766422216482402