schema: Move postgres special about TEXT/BYTEA to MWPlatform override
Global replaces on strings without boundaries or other checks could be
problematic, when the word is used as column name or table name.
This is not the case for mediawiki, because these names are provided in
lower case, but doing the override instead is clearer way to get this.

This also allows to define some special to get back the bytea if binary
support is needed in postgres.

Change-Id: I0a674408580d4e15222dc2556d597c2b523ce2b9