Skin: Skip LinkRenderer in footerLink and remove $wgFooterLinkCacheExpiry
Test plan:
* Enable $wgDebugToolbar.
* Load Main_Page and refresh once more.
* Apply this page and open another new tab to Main_page

Before:
> 16 queries, including:
> Skin::preloadExistence: `SELECT ..,page_is_redirect,..,page_len FROM
> page WHERE (page_namespace = 2 AND page_title = 'Admin')
> OR (page_namespace = 3 AND page_title = 'Admin')
> OR (page_namespace = 1 AND page_title = 'Main_Page')
> OR (page_namespace = 4 AND page_title IN
>  ('Privacy_policy','About','General_disclaimer')
> )`

After:
> 16 queries, including:
> Skin::preloadExistence: `SELECT ..,page_is_redirect,..,page_len FROM
> page WHERE (page_namespace = 2 AND page_title = 'Admin'
> OR (page_namespace = 3 AND page_title = 'Admin')
> OR (page_namespace = 1 AND page_title = 'Main_Page')`

The only other observable difference is that there is no longer a
title attribute on these links holding the destination page name.
This matches the other footer links such as
"By using this site, you agree to the Terms of Use and Privacy Policy."
and "Mobile view", "Developers", "Statistics" and "Cookie statement"
which also don't. If desired, this could be easily added though
by setting 'title' to the title page msg text value, the same used to
create the title object.

Bug: T313462
Change-Id: I581c49839ab4afe2542ee7d00393f11eff9879c6