rdbms: Remove LBFactory destructor
Follows-up I3adb57cbb1fd which added this in 2016 "just in case".

This has been causing fatal errors for several year since due to
LoadBalancer::closeAll calling into objects such as $this->logger,
which have in turn already began their own shutdown.

Destructors should generally be minimal and self-contained, which
these were not.

I could not find what case this is for, but I do note that the
Database objects have their own `__destruct` which calls
IDatabase::closeConnection, as implemented by DatabaseMysqli,
which means even if php-mysqli fails to disconnect, we should still
perform our own disconnect at the end of the process via destructors.

I've also removed the debug message while at it, as it appears to be
unreachable in practice and there are no explicit callers to closeAll.

Bug: T214734
Change-Id: I1ca2ad394fe39d444602d95e64e33408da3bce03