resourceloader: replace `currReqModules` with a boolean
Currently an array of the module names currently being requested,
but the only way it is used is to check if the array is empty. Just
use a boolean instead. Update documentation accordingly to remove
old reference to this array explaining why it wasn't used for
doRequest().

Follow-up: I00d6985c054fecd88acf73041aa02878e83d62bc
Change-Id: I3159a954f25bfc67934d0ee84035d7324ac21bdf