rdbms: gently discourage tableNames() and tableNamesN()
These methods should really not be used except in very special
circumstances. In particular, you definitely shouldn’t use them just
because you want to do a JOIN – we have much better methods for that
these days.

Also, fix the usage example for tableNames(). It returns an associative
array, which cannot be destructured into a sequential array [1]; since
PHP 7.1, you can destructure it into an associative array instead [2].
It was originally designed for use with the accursed (but still
supported) extract() function, which dumps the array keys into your
local scope, as shown in the comments added in commit 412eda0517:

    extract( $dbr->tableNames( 'user', 'watchlist' ) );
    // $user and $watchlist exist now

This was recognized as a bad pattern as early as 2006, when commit
a474761d9a introduced tableNamesN() to be used with list(). Apparently,
the tableNames() documentation was eventually changed to use list() as
well, without realizing that it doesn’t actually work this way.

[1]: https://3v4l.org/HGRKU
[2]: https://www.php.net/manual/en/migration71.new-features.php#migration71.new-features.support-for-keys-in-list

Change-Id: Ie870011e403034d8725d83e5185acc038f44249e