LogFormatter: Cast argument of ctype_digit to string [php8.1]
Passing ints here is deprecated in php 8.1. Additionally, it is
almost certainly not what we want, since an int to this function
is interpreted as an ascii code point (e.g. ctype_digit( 1 ) is
false. It is true for ints between 48-57 and > 255).

In practise this doesn't change much, since there is also an
is_int check here.

Bug: T313663
Change-Id: I11f3c710ba667d0f01574856af8f3da11d0b7f15