api-query-search: Request snippets when title variants are requested
Some search engines generate the title variant from the highlighted
snippet, and thus must have the snippet requested to be able to
provide the title. This is perhaps a bit implementation specific,
but until there is a more concrete api around requesting particular
values from the search engine this should work to provide users
a reasonable output, and document more of the limitations the
future api will need.

Bug: T306477
Change-Id: Ibcd8c24e87e424acab43ce842b68ebd5896e36cd