Move DOM transformations into HTMLTransformInput getters.
This moved the logic for applying page bundle data into the getters that
return DOM elements. It also makes the application of version
downgrades implicit.

NOTE: This patch changes the expected value of one of the phpunit tests
to a version that has no closing </div> tags. This appears to be the
original and expected behavior, per the corresponding test in the
parsoid extension's Parsoid.js test suite.

Change-Id: If2d7b06d8ba92fb63e6955ec7587ed4aea557251