resourceloader: optimize batchRequest() even further
Instead of having a dedicated variable to keep track of if there are
pending modules, just use `length`, which is needed to track the
length of the request URL anyway. Don't include the request base
length in `length`, so that length is truthy if and only if there is at
least one module pending, and just add it when needed in the
check for the URL becoming too long.

Follow-up: I3159a954f25bfc67934d0ee84035d7324ac21bdf
Change-Id: Id7e1f36bbb28a91f0cddf126b6e1a7b2346d8fa1