ResourceLoader: Remove outer TimeoutException try-catch in respond()
Follows-up I4c3770b9ef. While timeouts in load.php are extremely
rare (one in 90 days?), we did find one in the perf-team Logstash
dashboard and it coincided with this runtime error as result of the
added try-catch:

> PHP Notice: Undefined variable: etag

Remove the try-catch and let these fatal the same way as other very
early or very late errors with HTTP 500 instead.

Bug: T312806
Change-Id: Iffb238de2243d0885dcf1d78fcce7827b09cb84d