rdbms: Remove synchronous pre-send waitForReplication fallback
The `$workCallback` passed from MediaWiki.php to LBFactory is an
optional callback that Rdbms may invoke whenever it expects to idle
for some time before waiting for something.

In practice, this never happens because the main reason it was
introduced was for use in ChronologyProtector, and it was removed from
there last year in commit bd7cf4dce9 (I0456f5d40a).

The code wrapping for it has remained in-place, because there was
still one theoretical caller in LBFactory for when Redis (or some
other ChronologyProtector store) is down.

I propose that if Redis is down, the fallback is that we let
replication happen normally, which we expect in most cases to suffice
by itself, since replication is generally faster than a full
server-client-server RTT. CP logs a warning when this happens, for
operational visibility.

In the past 90 days, there are exactly 2 entries in unsampled WMF
Logstash for `channel:DBReplication message:persistSessionReplicationPositions`
(not including labswiki/cloudweb which were having problems in this
timeframe, failing for related reasons).

Logs for the two matching reqId:

1. 2022-07-14 wikidata POST /w/api.php (referer: -)
   > - MediumSpecificBagOStuff::doLock failed due to timeout for
   >   global:Wikimedia\Rdbms\ChronologyProtector:87f2d18…:v2.
   > - …::persistSessionReplicationPositions: failed to save primary
       positions for DB cluster(s) s8, cluster27

2. 2022-07-27 enwiki POST /w/index.php (referer: Special:CreateAccount)
   > - MediumSpecificBagOStuff::doLock failed due to timeout for
   >   global:Wikimedia\Rdbms\ChronologyProtector:d19e340…:v2.
   > - …::persistSessionReplicationPositions: failed to save primary
       positions for DB cluster(s) extension2

Bug: T314434
Change-Id: Id60cf994c7a40c50b487e6281c6bd114ddf13690