ParsoidHandler: do not emit etag for wt2html
Emitting a random ETag without storing it or the corresponding content
is not useful: The ETag cannot possibly ever match anything if the
client uses it in a later request with If-Match or If-None-Match.
If-Match will always fail, and If-None-Match will always succeed.

NOTE: When RESTbase proxies the response generated by this endpoint, it
will assign its own ETag to support stashing. It does not rely on the
ETag returned here.

Bug: T310710
Change-Id: I32b77a89549c37d32502adb101102747bc9ca45f