Unify the "magic variable" and "parser function" form of several built-ins
The following magic variables also have "parser function" forms, where
the first argument is a user-supplied title:

  pagename, pagenamee, fullpagename, fullpagenamee, subpagename,
  subpagenamee, rootpagename, rootpagenamee, basepagename,
  basepagenamee, talkpagename, talkpagenamee, subjectpagename,
  subjectpagenamee, pageid, cascadingsources, namespace, namespacee,
  namespacenumber, talkspace, talkspacee, subjectspace, subjectspacee

Refactor the code so that the magic variable form invokes the parser
function form with no arguments to reduce code duplication.  We also
tweak the behavior of parser function when invoked with no arguments,
although this change will not be directly visible because the parser
always prefers magic variables over parser functions, so the parser
function is never actually invoked with no arguments.

A future patch may allow the parser function to be invoked with a
hash prefix (I895087c546dc820c77c0dda596dfeb72586b87cc) in which case
consistency will be more important.

Note that `revisionuser`, `revisionid`, `revisionday`, `revisionday2`,
`revisionmonth`, `revisionmonth1`, `revisionyear` and
`revisiontimestamp` are also of a similar form and could be included
in this list, but their magic variable and parser function
implementations do not appear to be consistent.  This will be
addressed in future patches.

In addition, the following magic variables have a "parser function" form
where the presence or absence of the first argument selects "raw" output:

  numberofarticles, numberoffiles, numberofusers, numberofactiveusers,
  numberofpages, numberofadmins, numberofedits

Similar to above, refactor the code so that the magic variable form
invokes the parser function form with no arguments to reduce code
duplication (and to support future direct invocation of the parser
function with no arguments).

Bug: T204370
Change-Id: Iaec33fb40a2d9884daf2852ed6a6a3b53c9d3863