Set $wgServer for HttpRequestFactoryTest to prevent warnings on php8.1
If $wgServer is null, then wfParseUrl() will freak out, which will
cause warnings on php8.1

Note: This is not a backport, but a new commit to 1.38. Master
has refactored all of this code so the issue isn't present anymore,
but the refactoring seems rather large to backport. (472a914c63baa)

Bug: T314208
Change-Id: If920ed1bccf759e9f44958563ac2229c5e1cde32