rdbms: Stop accepting live connection in DBConnRef constructor
This just makes the class overly complex and defies the whole point of
having this class in the first place (=delaying getting db connections).

It makes it harder to read and understand and specially polymorphic
arguments are quite confusing.

It also would break the reloading db on the fly (I6c3ffde62f6e), users
of DBConnectionRef should not know about the underlying connection, this
class is trying encpsulate that and accepting connection in constructor
breaks this encapsulation.

Bug: T298485
Depends-On: I951ab99ae766788dbae8bd338aaae114a388b21f
Change-Id: I743c0565504c65c4dedf29f2a36d4c386601fda0