rdbms: Remove redundant error silencing for mysqlError()
This was introduced to deal with the fact that `mysql_error()`
could emit an error if passed a dead connection object.

This function is no longer called, and the DatabaseMysql class from
which it was removed from MW long ago (which existed to support the
php-mysql extension).

The newer php-mysqli extension and its $mysqli->error property
or mysqli_connect_error() method do not emit warnings. Thus this
is no longer needed.

Change-Id: If10f6535cdead7932a315e63478009fbae186ea8