rcfilters: Don't remove #contentSub on Special:Watchlist
Instead, only remove the elements inside of #contentSub that
we want to hide. The CSS classes for them were added recently
in b1ae47e4813ce858c68c8034e98586844d25463b.

It messed up the margins/padding on skins that display this element
and expect it to exist, and it didn't work on skins that did not
display it at all (the ID is not required to be there).

The CSS also applied on other pages, hiding e.g. the breadcrumbs
on Special:RecentChangesLinked. I think this was unintentional.

Bug: T227228
Change-Id: I87aefec80998db5a16cdbea56ca6bdf8a01ee317